feat: update ThemeProvider default theme to system; refactor GitHubIcon import and remove unused Montserrat font
