+++ src/ui/fonts.ts
@@ -1,8 +1,3 @@
-import { Inter, Montserrat } from "next/font/google";
+import { Inter } from "next/font/google";
 
 export const inter = Inter({ subsets: ["latin"], weight: ["400", "500", "600", "700"] });
-
-export const montserrat = Montserrat({
-  subsets: ["latin"],
-  weight: ["400", "500", "600", "700"],
-});
